requirements.txt
telepot
requests
python-dotenv
schedule